create generate csv method in filter_model.py

+++ tests/modules/test_filter.py
@@ -3,7 +3,6 @@ import os
 import pytest
 
 from echis.main import settings
-from echis.main.settings import ROOT_DIR
 from echis.modules.filter import is_bad_word
 
 
@@ -15,7 +14,7 @@ class TestFilter:
             writer.writerow(["fuck"])
         self.path = settings.FILTER
         yield
-        os.remove(self.path)
+        os.remove(self.path) if os.path.exists(self.path) else None
 
     def test_is_bad_word_get_true(self):
         is_bad = is_bad_word("fuck")
@@ -23,4 +22,9 @@ class TestFilter:
 
     def test_is_not_bad_word(self):
         is_bad = is_bad_word("random")
+        assert not is_bad
+
+    def test_file_not_exists_should_be_return_false(self):
+        os.remove(self.path)
+        is_bad = is_bad_word("fuck")
         assert not is_bad
